Update change assigned user urls
